Fetch Intercept: Collect mapping of tweets quoting other tweets
Since quoted tweets don't really have easily accessible links to scrape
the IDs from, we need to collect the mapping from parent tweets to the
quote tweets when intercepting data about media files. This will make
it possible to insert the download button for the video player inside
the quote.